Created separate class for DeDuplicating Long Strings kata
